requirements.txt
schedule
pystray
pillow
